Object revocation: generational region slots, retiring the lifetime cap
Untyped region names become generational (crates/slots), so destroy frees the slot
and the next create reuses it. Before this, region slots were never reused: the
256-entry table bounded region creations over the kernel's whole lifetime, not
concurrent use, so a system running workloads that come and go would eventually wedge
after 256 region creations ever. That is untenable for the "run real workloads" thesis
and is the last thing standing between object revocation and a repeatable spawn loop.

The change is mechanical: Object::Untyped(usize) -> Object::Untyped(u64), the
hand-rolled count-based array becomes a slots::Table<Region, 256>, create/split
insert_with, and destroy removes the slot (bumping its generation, so every stale
Untyped capability fails to resolve, the same stale-safety as Tids and endpoint names).
The type threads through cap.rs, syscall.rs, revoke.rs, user.rs, sched.rs, kmem.rs.

Now MAX_REGIONS bounds concurrent regions, not lifetime creations. The test proves it:
320 create+destroy cycles (past the old 256 cap) all succeed, one freed slot serving
the whole loop, frames netting to zero each iteration. All 112 tests pass.

+++ kernel/src/cap.rs
@@ -33,7 +33,7 @@ pub enum Object {
     /// **Untyped memory** (milestone 11): a capability to a chunk of raw physical memory the
     /// process may retype into pages. Invoking it grows the process's address space out of its
     /// own budget, and the kernel allocates nothing to do it. See kernel/src/untyped.rs.
-    Untyped(usize),
+    Untyped(u64),
 
     /// A hardware interrupt, by INTID.
     ///
@@ -142,7 +142,7 @@ pub fn irq_cap_rights(intid: u32, rights: Rights) -> Cap {
 }
 
 /// A capability to an untyped memory region. `WRITE` lets the holder retype pages out of it.
-pub fn untyped_cap(region: usize) -> Cap {
+pub fn untyped_cap(region: u64) -> Cap {
     Cap {
         object: Object::Untyped(region),
         rights: Rights::WRITE,
